fix: Make return type for FieldKind_INT an int64 (#1982)
## Relevant issue(s)

Resolves #1681 
Resolves #1981 

## Description

The PR ensures the return type for `FieldKind_INT` is always of the same
go type `int64`.

Fixing this also surface a small bug in the indexes where a broken index
should have returned an error but was returning an error for a different
reason. That reason being that the expected return type was `int64` but
the actual returned type was `uint64`.

+++ db/index.go
@@ -166,6 +166,13 @@ func (i *collectionSimpleIndex) Update(
 	if err != nil {
 		return err
 	}
+	exists, err := txn.Datastore().Has(ctx, key.ToDS())
+	if err != nil {
+		return err
+	}
+	if !exists {
+		return NewErrCorruptedIndex(i.desc.Name)
+	}
 	err = txn.Datastore().Delete(ctx, key.ToDS())
 	if err != nil {
 		return err
